refactor(world): split WorldServiceImpl into focused collaborators
Extract bulk import, the chat-input creation flow, and startup world
loading out of WorldServiceImpl into WorldImportCoordinator,
WorldCreationPrompts, and WorldLoadBootstrap. The service keeps its
public methods and delegates to the new collaborators; behaviour,
threading, and scheduling are preserved exactly.
