+++ buildsystem-core/src/main/java/de/eintosti/buildsystem/world/creation/WorldImportCoordinator.java
@@ -1,0 +1,188 @@
+/*
+ * Copyright (c) 2018-2026, Thomas Meaney
+ * Copyright (c) contributors
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+package de.eintosti.buildsystem.world.creation;
+
+import de.eintosti.buildsystem.BuildSystemPlugin;
+import de.eintosti.buildsystem.api.world.builder.Builder;
+import de.eintosti.buildsystem.api.world.creation.generator.Generator;
+import de.eintosti.buildsystem.api.world.data.BuildWorldType;
+import de.eintosti.buildsystem.storage.WorldStorageImpl;
+import de.eintosti.buildsystem.util.StringCleaner;
+import de.eintosti.buildsystem.world.WorldServiceImpl;
+import java.io.File;
+import java.util.Map;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.concurrent.atomic.AtomicInteger;
+import java.util.function.Predicate;
+import org.bukkit.Bukkit;
+import org.bukkit.entity.Player;
+import org.bukkit.scheduler.BukkitRunnable;
+import org.jspecify.annotations.NullMarked;
+import org.jspecify.annotations.Nullable;
+
+/**
+ * Coordinates staggered bulk imports of worlds, importing one world per configured delay interval while tracking
+ * whether an import is currently in progress.
+ */
+@NullMarked
+public class WorldImportCoordinator {
+
+    private final AtomicBoolean importingAllWorlds = new AtomicBoolean(false);
+
+    private final BuildSystemPlugin plugin;
+    private final WorldServiceImpl worldService;
+    private final WorldStorageImpl worldStorage;
+
+    public WorldImportCoordinator(
+            BuildSystemPlugin plugin, WorldServiceImpl worldService, WorldStorageImpl worldStorage) {
+        this.plugin = plugin;
+        this.worldService = worldService;
+        this.worldStorage = worldStorage;
+    }
+
+    public void importWorlds(Player player, String[] worldList, Generator generator, @Nullable Builder creator) {
+        int delay = plugin.getConfigService().current().world().importAllDelay();
+        plugin.getMessages()
+                .sendMessage(
+                        player, "worlds_importall_started", Map.entry("%amount%", String.valueOf(worldList.length)));
+        plugin.getMessages().sendMessage(player, "worlds_importall_delay", Map.entry("%delay%", String.valueOf(delay)));
+
+        importingAllWorlds.set(true);
+        BulkImportListener listener = new BulkImportListener() {
+            @Override
+            public void skippedExisting(String worldName) {
+                plugin.getMessages()
+                        .sendMessage(
+                                player, "worlds_importall_world_already_imported", Map.entry("%world%", worldName));
+            }
+
+            @Override
+            public void invalidName(String worldName, String invalidChar) {
+                plugin.getMessages()
+                        .sendMessage(
+                                player,
+                                "worlds_importall_invalid_character",
+                                Map.entry("%world%", worldName),
+                                Map.entry("%char%", invalidChar));
+            }
+
+            @Override
+            public void imported(String worldName) {
+                plugin.getMessages()
+                        .sendMessage(player, "worlds_importall_world_imported", Map.entry("%world%", worldName));
+            }
+        };
+        importStaggered(
+                        worldList,
+                        listener,
+                        worldName -> worldService.importWorld(
+                                player, worldName, creator, BuildWorldType.IMPORTED, generator, "void", false))
+                .thenRun(() -> plugin.getMessages().sendMessage(player, "worlds_importall_finished"));
+    }
+
+    public boolean isImportingAllWorlds() {
+        return importingAllWorlds.get();
+    }
+
+    public CompletableFuture<Integer> importWorlds() {
+        if (!importingAllWorlds.compareAndSet(false, true)) {
+            return CompletableFuture.failedFuture(new IllegalStateException("A bulk import is already in progress"));
+        }
+
+        String[] directories = scanImportableDirectories();
+        if (directories.length == 0) {
+            importingAllWorlds.set(false);
+            return CompletableFuture.completedFuture(0);
+        }
+
+        return importStaggered(
+                directories,
+                new BulkImportListener() {},
+                worldName -> worldService.importWorld(worldName).build() != null);
+    }
+
+    /**
+     * Per-world progress callbacks for a staggered bulk import.
+     */
+    private interface BulkImportListener {
+
+        default void skippedExisting(String worldName) {}
+
+        default void invalidName(String worldName, String invalidChar) {}
+
+        default void imported(String worldName) {}
+    }
+
+    /**
+     * Imports the given worlds one per configured delay interval, skipping names that are already imported or contain
+     * invalid characters. Clears {@link #importingAllWorlds} and completes with the import count once all names have
+     * been processed.
+     */
+    private CompletableFuture<Integer> importStaggered(
+            String[] worldNames, BulkImportListener listener, Predicate<String> importer) {
+        int delay = plugin.getConfigService().current().world().importAllDelay();
+        String invalidCharacters = plugin.getConfigService().current().world().invalidCharacters();
+
+        CompletableFuture<Integer> result = new CompletableFuture<>();
+        AtomicInteger index = new AtomicInteger(0);
+        AtomicInteger imported = new AtomicInteger(0);
+
+        new BukkitRunnable() {
+            @Override
+            public void run() {
+                int i = index.getAndIncrement();
+                if (i >= worldNames.length) {
+                    this.cancel();
+                    importingAllWorlds.set(false);
+                    result.complete(imported.get());
+                    return;
+                }
+
+                String worldName = worldNames[i];
+                if (worldStorage.worldExists(worldName)) {
+                    listener.skippedExisting(worldName);
+                    return;
+                }
+
+                String invalidChar = StringCleaner.firstInvalidChar(worldName, invalidCharacters);
+                if (invalidChar != null) {
+                    listener.invalidName(worldName, invalidChar);
+                    return;
+                }
+
+                if (importer.test(worldName)) {
+                    imported.incrementAndGet();
+                    listener.imported(worldName);
+                }
+            }
+        }.runTaskTimer(plugin, 0, 20L * delay);
+
+        return result;
+    }
+
+    private String[] scanImportableDirectories() {
+        String[] directories = Bukkit.getWorldContainer().list((dir, name) -> {
+            File worldFolder = new File(dir, name);
+            return worldFolder.isDirectory()
+                    && new File(worldFolder, "level.dat").exists()
+                    && !worldStorage.worldExists(name);
+        });
+        return directories != null ? directories : new String[0];
+    }
+}
